Refactor how UrlShortenerDatabaseService is constructed
After spending a bit of time considering how to test the class, it
seemed to me that, based on it extending TableGateway, that it would
just be too much work and that I'd be mocking things that were outside
the domain of the class. Based on that, this change refactors the class
to accept an AbstractTableGateway object as a constructor argument
instead of extending TableGateway. The class was also renamed
accordingly so that the name would be that much more meaningful.